rename protomodel immediate mapper
